Refactor Postgres class to simplify initialization and shutdown
Replaces `get_instance` singleton logic with direct initialization, adding default pool size parameters. Renames `close_pool` to `shutdown` and adjusts related logic to handle repeated calls safely. Updates test cases to align with these changes and improves maintainability.

+++ tests/test_postgres.py
@@ -1,5 +1,4 @@
 import datetime
-import json
 import random
 import unittest
 import uuid
@@ -8,10 +7,11 @@ import psycopg_pool
 import pydantic
 from psycopg import sql
 
-from pgraf import errors, postgres
+from pgraf import errors, postgres, utils
+from tests import common
 
 TEST_TABLE = """\
-CREATE TABLE public.test (
+CREATE TABLE IF NOT EXISTS public.test (
     id          UUID PRIMARY KEY,
     created_at  TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     value       INT NOT NULL,
@@ -27,10 +27,10 @@ INSERT INTO public.test (id, created_at, value, title, labels)
 """
 
 
-class TestModel(pydantic.BaseModel):
+class Model(pydantic.BaseModel):
     id: uuid.UUID = pydantic.Field(default_factory=uuid.uuid4)
     created_at: datetime.datetime = pydantic.Field(
-        default_factory=lambda: datetime.datetime.now(tz=datetime.UTC)
+        default_factory=utils.current_timestamp
     )
     value: int = pydantic.Field(default_factory=lambda: random.randint(0, 100))  # noqa: S311
     title: str | None = None
@@ -39,14 +39,15 @@ class TestModel(pydantic.BaseModel):
 
 class PostgresTestCase(unittest.IsolatedAsyncioTestCase):
     async def asyncSetUp(self) -> None:
-        self.postgres = postgres.Postgres()
+        self.postgres = postgres.Postgres(common.postgres_url())
         async with self.postgres.cursor() as cursor:
             await cursor.execute(TEST_TABLE)
 
     async def asyncTearDown(self) -> None:
-        async with self.postgres.cursor() as cursor:
-            await cursor.execute('DROP TABLE public.test')
-        await self.postgres.close_pool()
+        if self.postgres._pool:
+            async with self.postgres.cursor() as cursor:
+                await cursor.execute('DROP TABLE public.test')
+        await self.postgres.shutdown()
 
     async def test_pool_is_created(self) -> None:
         self.assertIsInstance(
@@ -59,9 +60,9 @@ class PostgresTestCase(unittest.IsolatedAsyncioTestCase):
         self.assertFalse(result)
 
     async def test_execute(self) -> None:
-        item = TestModel()
-        async with postgres.execute(
-            INSERT_TEST_RECORD, item.model_dump(), TestModel
+        item = Model()
+        async with self.postgres.execute(
+            INSERT_TEST_RECORD, item.model_dump(), Model
         ) as cursor:
             result = await cursor.fetchone()
         self.assertEqual(result.id, item.id)
@@ -69,18 +70,20 @@ class PostgresTestCase(unittest.IsolatedAsyncioTestCase):
         self.assertEqual(result.value, item.value)
 
     async def test_execute_with_sql_composable(self) -> None:
-        item = TestModel()
+        item = Model()
         composed_query = sql.Composed(
             [
-                sql.SQL('INSERT INTO'),
-                sql.Identifier('public.test'),
+                sql.SQL('INSERT INTO '),
+                sql.Identifier('public'),
+                sql.SQL('.'),
+                sql.Identifier('test'),
                 sql.SQL('(id, created_at, value)'),
-                sql.SQL('VALUES'),
+                sql.SQL(' VALUES '),
                 sql.SQL('(%(id)s, %(created_at)s, %(value)s) RETURNING *'),
             ]
         )
-        async with postgres.execute(
-            composed_query, item.model_dump(), TestModel
+        async with self.postgres.execute(
+            composed_query, item.model_dump(), Model
         ) as cursor:
             result = await cursor.fetchone()
         self.assertEqual(result.id, item.id)
@@ -89,7 +92,11 @@ class PostgresTestCase(unittest.IsolatedAsyncioTestCase):
 
     async def test_execute_with_database_error(self) -> None:
         with self.assertRaises(errors.DatabaseError):
-            async with postgres.execute(
+            async with self.postgres.execute(
                 'SELECT * FROM nonexistent_table', {}
             ) as cursor:
                 await cursor.fetchone()
+
+    async def test_shutdown_double_call(self) -> None:
+        await self.postgres.shutdown()
+        # Will be called again in teardown
